power-system: add DC-DC charger and temp. switches
This includes changes recommended to us during summer of 2025. Temporary
switch symbols have been used to indicate where to expect them, while
their internal circuit is not yet know and implemented. These will most
likely be replaced by some power transistors in the future.

A few control signals were added and are not yet connected to a control
board, but these shall be implemented on the power control PCB.