layout-2025: update some power symbols
New symbol for the charge controller now has two outputs: one for the
battery and one for the load. Also, its value was replaced with "TBD"
since we haven't decided on the actual model that will be used.

Some wires were replaced with power symbols to simplify the schematic.

A symbol was created for the cooling system and it was added to the
control box page. Since little is known about this system at the
moment, it is only represented by its power inputs. This power will
come from the main energy sources (fuel cells + auxiliary battery).

Information was added about the auxiliary battery model and capacity.
Many power symbols' fields were also uniformized to separate value,
part number, description and manufacturer.